LPS/backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1

